# Colorspace: improve performance