This patch series improves the performance (both memory usage and speed) of the color conversion code. I decided to look into this because some performance issues mention color conversion as a problematic step for performance. Primarily, this patch series prevents intermediate arrays from being created and improves clamping speed. The commit messages contain more information about the changes.

I'm including a benchmark of the changes, made using the steps from https://github.com/mozilla/pdf.js/wiki/Benchmarking-your-changes. The PDF files that I used are http://carto.metro.free.fr/documents/CartoMetroParis.v3.6.pdf (from #2541) and https://www.prorail.nl/sites/default/files/spoorkaart_prorail_april_2013.pdf (from #2813), each with 6 runs. Even though the speed-up is relatively small, these changes do result in lower rendering times (as confirmed by a second run of the benchmark). Creating less intermediate arrays also helps to reduce memory consumption somewhat, although I could not measure that reliably.

browser | stat         | Count | Baseline(ms) | Current(ms) | +/- |    %  | Result(P<.05)
------- | ------------ | ----- | ------------ | ----------- | --- | ----- | -------------
Firefox | Overall      |    12 |        12493 |       12421 | -72 | -0.58 |              
Firefox | Page Request |    12 |           28 |          32 |   4 | 14.46 |              
Firefox | Rendering    |    12 |        12465 |       12389 | -76 | -0.61 |