@timvandermeij 2016-12-02
I have updated the original description with a new benchmark that uses 25 runs per test case.