@brendandahl 2016-12-02
Reposting...It's been awhile since I looked at this, but using an inline conditional use to always be faster. If I remember correctly, we actually removed calls to clamp an inlined the clamping.