@timvandermeij 2016-12-06
I removed the commit with the clamping improvements as its performance impact is hard to measure objectively and it's unclear if the additional function call is bad or not (even though the code is more readable). This can be done later as it would otherwise unnecessarily delay landing of the other commits.

@Snuffleupagus Would you be willing to review this patch as well?